+++ NEWS.md
@@ -3,6 +3,7 @@ ClimaAnalysis.jl Release Notes
 
 v0.5.4
 ------
+- Added support for extraction dimension from functions, such as `times`.
 - Reorganized internal modules so that each file is a module.
 
 v0.5.3